Lint

+++ include/shadesmar/memory/dragons.h
@@ -11,6 +11,7 @@
 
 #include <immintrin.h>
 #include <thread>
+#include <vector>
 
 namespace shm::memory::dragons {
 static inline void *_rep_movsb(void *d, const void *s, size_t n) {
@@ -56,7 +57,7 @@ void *_multithread_avx_async_cpy(void *d, const void *s, size_t n) {
   auto *dVec = reinterpret_cast<__m256i *>(d);
   size_t nVec = n / sizeof(__m256i);
 
-  lldiv_t perWorker = div((long long)nVec, maxThreads);
+  lldiv_t perWorker = div((int64_t)nVec, maxThreads);
 
   size_t nextStart = 0;
   for (uint32_t threadIdx = 0; threadIdx < maxThreads; ++threadIdx) {
@@ -81,7 +82,7 @@ void *_multithreaded_memcpy(void *d, const void *s, size_t n) {
   std::vector<std::thread> threads;
   threads.reserve(maxThreads);
 
-  lldiv_t perWorker = div((long long)n, maxThreads);
+  lldiv_t perWorker = div((int64_t)n, maxThreads);
 
   size_t nextStart = 0;
   for (uint32_t threadIdx = 0; threadIdx < maxThreads; ++threadIdx) {
